refactor(events): collapse PermissionDenialReason into DeniedReasonKind
Refs #370 (item 4).

`PermissionDenialReason` (mcp/tools/approval.rs) and
`DeniedReasonKind` (dispatch/events.rs) had structurally identical
4-variant shapes joined by a `From` impl. The duplication couldn't
actually diverge — adding a variant on one side broke the `From`
impl at compile time, forcing the other to change too — so the
"separation of concerns" justification was hollow.

Move `PermissionDenialReason::message()` onto `DeniedReasonKind`
(events module owns the canonical type). Replace the duplicate enum
in `mcp/tools/approval.rs` with a `pub use` re-export under the old
name, so existing call sites keep their `PermissionDenialReason::Variant`
syntax. Drop the now-unused `From` impl. `record_permission_denied`
no longer needs `reason_kind.into()` — the field type now matches.

Net: -52 lines, one canonical enum, dependency direction is now
mcp::tools → dispatch::events (the natural layering).

15/15 permission_prompt tests pass unchanged; no behavioral or
on-disk shape change.
